backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "channels-redis>=4.2.1",
    "django>=5.2.3",
    "django-channels>=0.7.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "pillow>=11.2.1",
    "python-decouple>=3.8",
    "redis>=6.2.0",
]
